[STP-773] Fix print question by id in Excel table. Change summary type to double

Approved-by: Yevgeny Vasilyev

+++ stp-domain/src/main/java/com/sytoss/domain/bom/personalexam/PersonalExamReportModel.java
@@ -22,7 +22,7 @@ public class PersonalExamReportModel {
 
     private String email;
 
-    private Float summary;
+    private Double summary;
 
     private PersonalExamStatus personalExamStatus;
